refactor: add candle as an imlplemented training backend
