+++ src/kettle/training/candle/__init__.py
@@ -1,0 +1,9 @@
+"""
+Candle backend for attestable training.
+
+This module provides the Candle ML framework backend implementation.
+"""
+
+from .tool import CandleTrainingTool
+
+__all__ = ["CandleTrainingTool"]
