+++ src/kettle/training/candle/src/model.rs
